refactor content out of components
