+++ README.md
@@ -1,6 +1,6 @@
-# GTA SA savegame editor documentation [![GitHub Pages Badge](https://img.shields.io/badge/docs-gh--pages-lightgrey.svg)](https://gtasa-savegame-editor.github.io/docs/#/) [![TravisCI Badge](https://api.travis-ci.org/gtasa-savegame-editor/docs.svg?branch=master)](https://travis-ci.org/gtasa-savegame-editor/docs)
+# GTA:SA Savegame Editor documentation [![GitHub Pages Badge](https://img.shields.io/badge/docs-gh--pages-lightgrey.svg)](https://gtasa-savegame-editor.github.io/docs/#/) [![TravisCI Badge](https://api.travis-ci.org/gtasa-savegame-editor/docs.svg?branch=master)](https://travis-ci.org/gtasa-savegame-editor/docs)
 
-## Dev
+## Usage
 
 Install Dependencies:
 
@@ -9,9 +9,3 @@ Install Dependencies:
 Start dev server:
          
     $ npm start
-
-## Deploy
-
-Deploy to GitHub Pages:
-
-    $ npm run deploy
