+++ src/content/blocks/01.js
@@ -1,25 +1,25 @@
 import React, { Component } from 'react';
-import icon05 from '../../../icon/05.gif';
-import icon08 from '../../../icon/08.gif';
-import icon0C from '../../../icon/0C.gif';
-import icon0D from '../../../icon/0D.gif';
-import icon0F from '../../../icon/0F.gif';
-import icon10 from '../../../icon/10.gif';
-import icon17 from '../../../icon/17.gif';
-import icon18 from '../../../icon/18.gif';
-import icon19 from '../../../icon/19.gif';
-import icon1C from '../../../icon/1C.gif';
-import icon22 from '../../../icon/22.gif';
-import icon24 from '../../../icon/24.gif';
-import icon25 from '../../../icon/25.gif';
-import icon26 from '../../../icon/26.gif';
-import icon28 from '../../../icon/28.gif';
-import icon2A from '../../../icon/2A.gif';
-import icon2B from '../../../icon/2B.gif';
-import icon2C from '../../../icon/2C.gif';
-import icon2E from '../../../icon/2E.gif';
-import icon2F from '../../../icon/2F.gif';
-import icon34 from '../../../icon/34.gif';
+import icon05 from '../../icon/05.gif';
+import icon08 from '../../icon/08.gif';
+import icon0C from '../../icon/0C.gif';
+import icon0D from '../../icon/0D.gif';
+import icon0F from '../../icon/0F.gif';
+import icon10 from '../../icon/10.gif';
+import icon17 from '../../icon/17.gif';
+import icon18 from '../../icon/18.gif';
+import icon19 from '../../icon/19.gif';
+import icon1C from '../../icon/1C.gif';
+import icon22 from '../../icon/22.gif';
+import icon24 from '../../icon/24.gif';
+import icon25 from '../../icon/25.gif';
+import icon26 from '../../icon/26.gif';
+import icon28 from '../../icon/28.gif';
+import icon2A from '../../icon/2A.gif';
+import icon2B from '../../icon/2B.gif';
+import icon2C from '../../icon/2C.gif';
+import icon2E from '../../icon/2E.gif';
+import icon2F from '../../icon/2F.gif';
+import icon34 from '../../icon/34.gif';
 import { Link } from 'react-router-dom';
 class Block01 extends Component {
   render() {
