+++ src/content/legend.js
