+++ src/content/pickups.js
