+++ src/content/sprites.js
@@ -1,66 +1,66 @@
 import React, { Component } from 'react';
-import icon01 from '../../icon/01.gif';
-import icon02 from '../../icon/02.gif';
-import icon04 from '../../icon/04.gif';
-import icon05 from '../../icon/05.gif';
-import icon06 from '../../icon/06.gif';
-import icon07 from '../../icon/07.gif';
-import icon08 from '../../icon/08.gif';
-import icon09 from '../../icon/09.gif';
-import icon0a from '../../icon/0A.gif';
-import icon0b from '../../icon/0B.gif';
-import icon0c from '../../icon/0C.gif';
-import icon0d from '../../icon/0D.gif';
-import icon0e from '../../icon/0E.gif';
-import icon0f from '../../icon/0F.gif';
-import icon10 from '../../icon/10.gif';
-import icon11 from '../../icon/11.gif';
-import icon12 from '../../icon/12.gif';
-import icon13 from '../../icon/13.gif';
-import icon14 from '../../icon/14.gif';
-import icon15 from '../../icon/15.gif';
-import icon16 from '../../icon/16.gif';
-import icon17 from '../../icon/17.gif';
-import icon18 from '../../icon/18.gif';
-import icon19 from '../../icon/19.gif';
-import icon1a from '../../icon/1A.gif';
-import icon1b from '../../icon/1B.gif';
-import icon1c from '../../icon/1C.gif';
-import icon1d from '../../icon/1D.gif';
-import icon1e from '../../icon/1E.gif';
-import icon1f from '../../icon/1F.gif';
-import icon20 from '../../icon/20.gif';
-import icon21 from '../../icon/21.gif';
-import icon22 from '../../icon/22.gif';
-import icon23 from '../../icon/23.gif';
-import icon24 from '../../icon/24.gif';
-import icon25 from '../../icon/25.gif';
-import icon26 from '../../icon/26.gif';
-import icon27 from '../../icon/27.gif';
-import icon28 from '../../icon/28.gif';
-import icon29 from '../../icon/29.gif';
-import icon2a from '../../icon/2A.gif';
-import icon2b from '../../icon/2B.gif';
-import icon2c from '../../icon/2C.gif';
-import icon2d from '../../icon/2D.gif';
-import icon2e from '../../icon/2E.gif';
-import icon2f from '../../icon/2F.gif';
-import icon30 from '../../icon/30.gif';
-import icon31 from '../../icon/31.gif';
-import icon32 from '../../icon/32.gif';
-import icon33 from '../../icon/33.gif';
-import icon34 from '../../icon/34.gif';
-import icon35 from '../../icon/35.gif';
-import icon36 from '../../icon/36.gif';
-import icon37 from '../../icon/37.gif';
-import icon38 from '../../icon/38.gif';
-import icon39 from '../../icon/39.gif';
-import icon3a from '../../icon/3A.gif';
-import icon3b from '../../icon/3B.gif';
-import icon3c from '../../icon/3C.gif';
-import icon3d from '../../icon/3D.gif';
-import icon3e from '../../icon/3E.gif';
-import icon3f from '../../icon/3F.gif';
+import icon01 from '../icon/01.gif';
+import icon02 from '../icon/02.gif';
+import icon04 from '../icon/04.gif';
+import icon05 from '../icon/05.gif';
+import icon06 from '../icon/06.gif';
+import icon07 from '../icon/07.gif';
+import icon08 from '../icon/08.gif';
+import icon09 from '../icon/09.gif';
+import icon0a from '../icon/0A.gif';
+import icon0b from '../icon/0B.gif';
+import icon0c from '../icon/0C.gif';
+import icon0d from '../icon/0D.gif';
+import icon0e from '../icon/0E.gif';
+import icon0f from '../icon/0F.gif';
+import icon10 from '../icon/10.gif';
+import icon11 from '../icon/11.gif';
+import icon12 from '../icon/12.gif';
+import icon13 from '../icon/13.gif';
+import icon14 from '../icon/14.gif';
+import icon15 from '../icon/15.gif';
+import icon16 from '../icon/16.gif';
+import icon17 from '../icon/17.gif';
+import icon18 from '../icon/18.gif';
+import icon19 from '../icon/19.gif';
+import icon1a from '../icon/1A.gif';
+import icon1b from '../icon/1B.gif';
+import icon1c from '../icon/1C.gif';
+import icon1d from '../icon/1D.gif';
+import icon1e from '../icon/1E.gif';
+import icon1f from '../icon/1F.gif';
+import icon20 from '../icon/20.gif';
+import icon21 from '../icon/21.gif';
+import icon22 from '../icon/22.gif';
+import icon23 from '../icon/23.gif';
+import icon24 from '../icon/24.gif';
+import icon25 from '../icon/25.gif';
+import icon26 from '../icon/26.gif';
+import icon27 from '../icon/27.gif';
+import icon28 from '../icon/28.gif';
+import icon29 from '../icon/29.gif';
+import icon2a from '../icon/2A.gif';
+import icon2b from '../icon/2B.gif';
+import icon2c from '../icon/2C.gif';
+import icon2d from '../icon/2D.gif';
+import icon2e from '../icon/2E.gif';
+import icon2f from '../icon/2F.gif';
+import icon30 from '../icon/30.gif';
+import icon31 from '../icon/31.gif';
+import icon32 from '../icon/32.gif';
+import icon33 from '../icon/33.gif';
+import icon34 from '../icon/34.gif';
+import icon35 from '../icon/35.gif';
+import icon36 from '../icon/36.gif';
+import icon37 from '../icon/37.gif';
+import icon38 from '../icon/38.gif';
+import icon39 from '../icon/39.gif';
+import icon3a from '../icon/3A.gif';
+import icon3b from '../icon/3B.gif';
+import icon3c from '../icon/3C.gif';
+import icon3d from '../icon/3D.gif';
+import icon3e from '../icon/3E.gif';
+import icon3f from '../icon/3F.gif';
 
 class Sprites extends Component {
   render() {
