+++ src/content/vehicle_colors.js
