+++ src/content/vehicle_types.js
