+++ src/content/weapons.js
